+++ actix-rt/CHANGES.md
@@ -1,5 +1,14 @@
 # Changes
 
+## [0.2.1] - 2019-03-11
+
+### Added
+
+* Arbiter::exec_fn - execute fn on the arbiter's thread
+
+* Arbiter::exec - execute fn on the arbiter's thread and wait result
+
+
 ## [0.2.0] - 2019-03-06
 
 * `run` method returns `io::Result<()>`